+++ src/InvestorApi.Domain/Entities/Position.cs
@@ -10,7 +10,7 @@ namespace InvestorApi.Domain.Entities
             // Required for instantiation by repository.
         }
 
-        private Position(Guid id, Guid accountId, string symbol, int quantity, decimal averagePrice)
+        private Position(Guid id, Guid accountId, string symbol, long quantity, decimal averagePrice)
         {
             Id = id;
             AccountId = accountId;
@@ -25,23 +25,23 @@ namespace InvestorApi.Domain.Entities
 
         public string Symbol { get; private set; }
 
-        public int Quantity { get; private set; }
+        public long Quantity { get; private set; }
 
         public decimal AveragePrice { get; private set; }
 
-        public static Position Create(Account account, string symbol, int quantity, decimal averagePrice)
+        public static Position Create(Account account, string symbol, long quantity, decimal averagePrice)
         {
             return new Position(Guid.NewGuid(), account.Id, symbol, quantity, averagePrice);
         }
 
-        public void Buy(int additionalQuantity, decimal price, decimal brokerageFees)
+        public void Buy(long additionalQuantity, decimal price, decimal brokerageFees)
         {
             // Calculate the new average price.
             AveragePrice = ((Quantity * AveragePrice) + (additionalQuantity * price) + brokerageFees) / (Quantity + additionalQuantity);
             Quantity += additionalQuantity;
         }
 
-        public void Sell(int quantity)
+        public void Sell(long quantity)
         {
             Quantity -= quantity;
         }
