+++ src/InvestorApi.Domain/Services/AccountService.cs
@@ -136,7 +136,7 @@ namespace InvestorApi.Domain.Services
         /// <param name="symbol">The share symbol.</param>
         /// <param name="quantity">The quantity to buy.</param>
         /// <param name="nonce">The nonce value required to detect suplicate orders.</param>
-        public void BuySharesAtMarketPrice(Guid userId, Guid accountId, string symbol, int quantity, long nonce)
+        public void BuySharesAtMarketPrice(Guid userId, Guid accountId, string symbol, long quantity, long nonce)
         {
             Quote quote = _shareQuoteProvider.GetQuote(symbol);
             if (quote == null)
@@ -159,7 +159,7 @@ namespace InvestorApi.Domain.Services
         /// <param name="symbol">The share symbol.</param>
         /// <param name="quantity">The quantity to sell.</param>
         /// <param name="nonce">The nonce value required to detect suplicate orders.</param>
-        public void SellSharesAtMarketPrice(Guid userId, Guid accountId, string symbol, int quantity, long nonce)
+        public void SellSharesAtMarketPrice(Guid userId, Guid accountId, string symbol, long quantity, long nonce)
         {
             Quote quote = _shareQuoteProvider.GetQuote(symbol);
             if (quote == null)
